requirements.txt
click==6.7
Flask==0.12.1
get==0.0.0
itsdangerous==0.24
Jinja2==2.9.6
lxml==3.7.3
MarkupSafe==1.0
post==0.0.0
public==0.0.0
query-string==0.0.0
request==0.0.0
setupfiles==0.0.0
Werkzeug==0.12.1
